core: improve disposisi

+++ surat/index.php
@@ -9,10 +9,12 @@
 $sql = "SELECT surat.id as ide, nomor_surat, tanggal_surat, pengirim, penerima, nomor_agenda, tanggal_agenda, buku_id, buku.nama, status, tipe_surat FROM surat 
         INNER JOIN buku ON surat.buku_id=buku.id WHERE tipe_surat = 'masuk' ORDER BY ide DESC";
 $result = mysqli_query($conn, $sql);
+$count = mysqli_num_rows($result);
 
 $sql2 = "SELECT surat.id as ide, nomor_surat, tanggal_surat, pengirim, penerima, nomor_agenda, tanggal_agenda, buku_id, buku.nama, status, tipe_surat FROM surat 
         INNER JOIN buku ON surat.buku_id=buku.id WHERE tipe_surat = 'keluar' ORDER BY ide DESC";
 $result2 = mysqli_query($conn, $sql2);
+$count2 = mysqli_num_rows($result2);
 ?>
 
 <!-- Begin Page Content -->
@@ -60,6 +62,7 @@ $result2 = mysqli_query($conn, $sql2);
 
                                     <tbody>
 
+                                    <?php if($count > 0) { ?>
                                         <?php foreach ($result as $key => $data) : ?>
                                             <tr>
                                                 <td><?php echo $key + 1 ?></td>
@@ -81,6 +84,13 @@ $result2 = mysqli_query($conn, $sql2);
                                                 <td><?php echo $data['tipe_surat'] ?></td>
                                             </tr>
                                         <?php endforeach ?>
+                                        <?php } else { ?>
+                                            <tr>
+                                                <td colspan="11"><b>
+                                                        <center>Tidak ada data</center>
+                                                    </b></td>
+                                            </tr>
+                                        <?php } ?>
 
                                     </tbody>
 
@@ -112,27 +122,35 @@ $result2 = mysqli_query($conn, $sql2);
 
                                     <tbody>
 
-                                        <?php foreach ($result2 as $key => $data) : ?>
+                                        <?php if ($count2 > 0) { ?>
+                                            <?php foreach ($result2 as $key => $data) : ?>
+                                                <tr>
+                                                    <td><?php echo $key + 1 ?></td>
+                                                    <td>
+                                                        <div class="d-flex">
+                                                            <a href="detail.php?id=<?php echo $data['ide'] ?>" class="btn btn-secondary mr-2"> Detail </a>
+                                                            <a href="edit.php?id=<?php echo $data['ide'] ?>" class="btn btn-warning mr-2"> Edit </a>
+                                                            <a href="proses_hapus.php?id=<?php echo $data['ide'] ?>" class="btn btn-danger"> Hapus </a>
+                                                        </div>
+                                                    </td>
+                                                    <td><?php echo $data['nomor_surat'] ?></td>
+                                                    <td><?php echo $data['tanggal_surat'] ?></td>
+                                                    <td><?php echo $data['pengirim'] ?></td>
+                                                    <td><?php echo $data['penerima'] ?></td>
+                                                    <td><?php echo $data['nomor_agenda'] ?></td>
+                                                    <td><?php echo $data['tanggal_agenda'] ?></td>
+                                                    <td><?php echo $data['nama'] ?></td>
+                                                    <td><?php echo $data['status'] ?></td>
+                                                    <td><?php echo $data['tipe_surat'] ?></td>
+                                                </tr>
+                                            <?php endforeach ?>
+                                        <?php } else { ?>
                                             <tr>
-                                                <td><?php echo $key + 1 ?></td>
-                                                <td>
-                                                    <div class="d-flex">
-                                                        <a href="detail.php?id=<?php echo $data['ide'] ?>" class="btn btn-secondary mr-2"> Detail </a>
-                                                        <a href="edit.php?id=<?php echo $data['ide'] ?>" class="btn btn-warning mr-2"> Edit </a>
-                                                        <a href="proses_hapus.php?id=<?php echo $data['ide'] ?>" class="btn btn-danger"> Hapus </a>
-                                                    </div>
-                                                </td>
-                                                <td><?php echo $data['nomor_surat'] ?></td>
-                                                <td><?php echo $data['tanggal_surat'] ?></td>
-                                                <td><?php echo $data['pengirim'] ?></td>
-                                                <td><?php echo $data['penerima'] ?></td>
-                                                <td><?php echo $data['nomor_agenda'] ?></td>
-                                                <td><?php echo $data['tanggal_agenda'] ?></td>
-                                                <td><?php echo $data['nama'] ?></td>
-                                                <td><?php echo $data['status'] ?></td>
-                                                <td><?php echo $data['tipe_surat'] ?></td>
+                                                <td colspan="11"><b>
+                                                        <center>Tidak ada data</center>
+                                                    </b></td>
                                             </tr>
-                                        <?php endforeach ?>
+                                        <?php } ?>
 
                                     </tbody>
